backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.9.2
scikit-learn==1.5.2
numpy==2.1.3
openai==1.54.3
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
pypdf==4.3.1
SQLAlchemy==2.0.36
